[WebApplication] - Refactored Application configuration files into config package
