+++ Server/online.quar/src/main/java/online/quar/application/config/MvcConfig.java
@@ -1,4 +1,4 @@
-package online.quar.application;
+package online.quar.application.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
@@ -14,6 +14,9 @@ public class MvcConfig implements WebMvcConfigurer {
         registry.addViewController("/hello").setViewName("pages/hello");
         registry.addViewController("/login").setViewName("pages/login");
         registry.addViewController("/carController").setViewName("pages/carController");
+        registry.addViewController("/webSocketOrgTest").setViewName("pages/webSocketOrgTest");
+        registry.addViewController("/webSocketTest").setViewName("pages/webSocketTest");
+
     }
 
 }
