+++ Server/online.quar/src/main/java/online/quar/application/config/ServerConfig.java
@@ -1,4 +1,4 @@
-package online.quar.application;
+package online.quar.application.config;
 
 import org.apache.catalina.Context;
 import org.apache.catalina.connector.Connector;
